CFID-293: Add logout.do to UI filter chain.
Also set default target for login to be "/" and for logout to
the "/login". This means that when a user logs out they will be
sent to the login page, rather than the previous behaviour
where they were redirected to the protected "/", causing a
session creation to cache the request, followed by a redirect
to the login page.

Change-Id: I5a558f31e3891612b2b5e13db2fd4b7e1199cd20