remove default jwt signing key
This might break unit/integration tests.  I'm mainly putting this up
for Tim, as the uaa that they use ends up picking up the value
removed in this commit.  UAA guys, feel free to -2 this and do it however
you want, but this needs to not override the value from config/uaa.yml.

Change-Id: I8a542cedc5010e9ff9cfc594832454408167f3f8