CFID-402: /Groups/{group}/Users endpoint
A special purpose endpoint for use as a user id/name
translation api.  It will be used by vmc so it has
to be quite restricted in function (i.e. it's not
a general purpose groups or users endpoint).  You
can only query it, as a user, for usernames and user
ids in the same group as you.  Otherwise the API
is the same as /Users.

[Fixes #35352703]

Change-Id: Ia592eb3bcf77b8afaef7e7b1f6bf9a3bc548ce62