[cfid-71] Add CORS filter to api sample
The issue is not to do with the login server specificially,
just that I was testing it with the api service in a different
domain than the app sample, so the browser common-origin
policy was preventing the xhr from completing (and giving very
little clue as to what was happening other than a pre-flight
OPTIONS request).

I added a filter to the api sample that I thought would fix it
but it hasn't.  Not really a super high priority to fix though.

[#35832353]: Fix implicit demo in app sample

Change-Id: I780786c60723fa0eb13e0a69c89dc75c9e6aaafe