[#36678769] Add tab-completion support for uaac
uaac-completion.sh is a bash script to be sourced for tab-completion to work.
It uses a completion-helper ruby script to generate the completion options based on the
command seen so far, eg.
	input='uaac secret', 		output='set change'
	input='uaac secret set',	output='--secret -s'

The completion-helper ruby script is a generic script that could potentially be used for
generating completion options for ANY command (bosh, vmc etc.)

Change-Id: Ib14509b45db3012fb3636eaf885b0af5fc4e240c