Return code for default vmc should be 302
The problem turned out to be that I was running tests locally
against a postgres database that had a non-default vmc registration.
In this database vmc had a registered callback (by default it does
not), and so the response was 401 because the redirect sent by the
test did not match.

[Fixes #36770647] UAA login authentication integration test fails (302 instead of 401)

Change-Id: I6774716660ba3b0063fc62a826ed8c124eac1325