[cfid-144] Fix race condition when users are deleted during a filter operation
If users are deleted in between the query being run and the last page being
iterated you can get a NoSuchElementException.  I fixed it by wrapping
the iterator and catching the exception. The result is that the size of
a list might be greater than the number of elements in its iterator
but there shouldn't be any more errors.

[#36825399] [cfid-144] scim query fails when over 200 users and no pagination parameters

Change-Id: If2a62b28d815f463125c1bd11b77056706ad9755