[cfid-154] Make sure Views are not cached by ForwardAware*
Caching view instances is done by default in Spring MVC by name
whereas we sometimes use teh same view name to refer to different
content types. The problem manifests itself when a user renders
the approval page in a browser (HTML) before the login server
has a chance to grab it in JSON.  If only the login server
accesses that page everything is fine (in once people stop
using the UI at UAA it would be fine, but only if everyone stops
forever).  Also relevant to [#37014635] [cfid-153].

[Fixes #37022525] [cfid-154] UAA is not returning options and
scopes in JSON from /authorize

Change-Id: I50ab36f8085d41db64034adcff9857ee28b6c74f