[cfid-157] Populate UaaUser.authorities from group_membership table instead of authorities field.
Requests to /oauth/authorize load a UaaUser object by directly accessing the users table.
After the addition of Scim Groups feature, the authorities field in the users table is
no longer kept in sync with the user's membership data. Hence, we get this info from the
group tables instead.

Fixes [#37075077].

Change-Id: I28b4beccc94ae87cf4fe3c389e497ac9bc27e4be