[cfid-232] Fix scope binding problems on authorize endpoint
Mostly this problem manifested itself in the login sample, but
a more exhaustive search revealed corner cases in the /authorize
endpoint itself (hence the bumped SECOAUTH dependency).  The issue
arises when scope=... is posted back to /authorize during the
approval process (it shouldn't be posted, but if it is this problem
arises).

[Fixes #38436243] [cfid-232] Login sample mangles scopes if specified by client

Change-Id: Ieb1757ed674d5bc7d970481ee08e887afd458c72