[cfid-124] Redirect to HTTPS if enabled
[#36349795]. Adds a Spring bean post-processor which inserts an additional
filter at the start of each Spring Security filter chain. If
the uaa.require_https property is set, the filter will check whether
the incoming request is secure or not. Depending on the filter chain, it
will either redirect (with a 301) or send an error. The latter is the default.
Filter chains for which a redirect should be performed can be specified by
name. The functionality relies on the accuracy of the servlet request isSecure
method. It shouldn't be enabled unless tomcat is configured with the
RemoteIpValve to check whether the appropriate header is present to indicate
that the request was over https (typically x-forwarded-proto).

This change also adds support for setting the HSTS (strict transport) header
on the response to tell browsers only to use HTTPS for the site.

The extra filter can also be configured to dump additional data on the
request being handled by a filter chain. If uaa.dump_requests is set in
combination with debug level logging, the complete set of request headers will
be logged per request to aid debugging.

Change-Id: I5f2178791a45cb1bf36d13bb2e01582fdf81d566