[cfid-348] Refactor app sample to use UAA scopes
This change doesn't affect the server (provider) features of the UAA
but it does alter the way that the sample client works, and provides
additional library features to support that.  The basic idea is
that a client app can interrogate the token scopes and use them to
provide role-based (or other) authorization locally. The sample does
this with the /check_token endpoint (hence the app has to be a
uaa.resource), but it could in principle work with local decoding too.

Also inspired by problems encountered with a abug that surfaced in the
dashboard:

[#39194235] [cfid-300] User authenticated by Login Server only has restricted authorities (uaa.user)

Change-Id: I01e760fbe7b9689263af61a2165f8b0432342ac1