[cfid-331] Prevent cycles in nested groups from causing an infinite recursion
Preventing formation of cycles in the group hierarchy during create or update will be
tricky (multiple points in code where we will have to do this), and also expensive
compared to what it will gain us in exchange. Detecting an existing cycle on the other
hand is fairly straightforward AND contained to one flow - GET /User/{id} where we
recursively compute the groups that a user is part of.

Fixes [#39895111].

Change-Id: I3516b2000a42e704918379ce13af6b68f37f0d48