Persist user authorizations
User authorizations are now persisted. The user authorization prompts
are now displayed only when a scope is requested that has not already
been approved or when that approval expires.

NOTE: This is not an ideal solution. That solution includes modifying
the AuthorizationEndpoint from SECOAUTH to handle this.

TODO: Prompt the user with only the scopes that need approval
TODO: Expand the design to include approved as well as denied scopes

Change-Id: Idbd55197e6e82a27ca8477c5db8864e1d4970bab