[cfid-256] Implement group member roles as a legitimate SCIM extension.
The SCIM spec allows for extension to the core schema in the form of new top-level
attributes, but not sub-attributes for existing top-level attributes. This makes the
'authorities' sub-attribute of the SCIM group's 'members' attribute a spec violation.

This change [#38709619] makes the SCIM Groups API spec compliant by removing the
authorities sub-attribute and instead adding top-level attributes for different type of
group members, like 'readers', 'writers' etc.

In addition, this change partially addresses [cfid-48][#35352689] - by having the custom
JSON deserializer for ScimGroup treat field names in a case-insensitive manner.

Change-Id: Icabb46ab2a2dcb3d3913bd6c48549e730d22a314