[cfid-598] remove small uaa.yml from war file
All the entries from the uaa.yml have been moved into inline
XML.  The file is still there with a comment to say how to override
it.  The main problem that we might anticipate is that there is
no longer an admin client in the bootstrap except in the default
profile.  Most of our users seem to want to put the admin client
in ther custom uaa.yml anyway, so that's probably not such a
problem.

[Fixes #44222845]