[cfid-627] Partial fix for infinite loop in implicit sample app
If the "app" client does not have implicit grant type there is an error
and the jso library cannot handle it.  The library is obviously not very
useful, so maybe we can switch to Scott's s2js library soon.  In the meantime
just adding an implicit grant type to the client (in default profile) should
work.

[Fixes #44579465] [cfid-627] clicking implicit demo on the sample app takes it into an infinite loop