Validate YAML with domain knowledge
Added test case using JSR-303 and Spring friends to validate
an Object bound to a Map.  The Map can come from uaa.yml and
the BindErrors will tell the user what is wrong.  Needs to be
expanded into a bean that sits in Spring config and is injected
with the @config map.

[#46077585]