Add the password reset endpoint
users are retrieved by email, and their user id is embedded in the expiring code

In cloud foundry, the user's email matches their username. By using the email field, we avoid the need to validate the username as an email before attempting to send to it.  This also maintains consistency with the forgot password page UI, for other uaa deployments