Github Issue https://github.com/cloudfoundry/uaa/issues/71
We should be able to handle users that are uaa.admin

UaaAuthentication stores it the same way that the client auth does. We can directly check the Authentication.getAuthorities for the required scope. Oauth2Authentication stores the scopes a bit differently, so we check it the same way as we do in our XML <http> patterns by leveraging the Spring Oauth2 expression handler.