Implement a revocable token strategy for stateless JWT token Goal: A token should be revoked if 1. A revoke endpoint is invoked with a user ID (revoke all tokens for that user) 2. A revoke endpoint is invoked with a client ID (revoke all the tokens for that client) 3. A client secret changes 4. A user password changes 5. A user username changes 6. A user email changes
Solution:
Both client and user have a salt field to satisfy 1. and 2.
Upon changing the salt, a validation of an existing token will fail
Use client_secret, username, email and password as for of the revocation signature. Any changes to these
fields will cause the validation to fail

All tokens are from now on revocable, with existing tokens backwards compatible
Use a murmur hash to shorten the hash value, rather than using SHA or MD algorithms

Ensure that CID is used (and that there is a corresponding CLIENT_ID)