oidc: set JWT iss claim to UAA server location
Right now, the iss claim in a JWT points to the OAuth endpoint used to
get the token.  This makes OIDC integration impossible because the iss
claim must match the issuer property of the OIDC configuration document.
With this change, we no longer append "oauth/token" to the issuer URL.