Only use HTTPS maven repositories
It has been known for a long time that its important to use HTTPS for maven repositories to avoid attacks on developers, build systems, and end users.

This commit updates repos to use HTTPS and more up to date spring.io domains.

It also removes the objectstyle.org maven repo which doesn't seem to be used, is http-only, and seems likely to go away eventually [1]

[1] http://www.objectstyle.com/objectstyle-org-and-community-roots