oidc: Use standard JWA algorithm names in /token_key(s)
The JWA specification has a list of algorithms and the registered "alg"
name to be used in the /token_key(s) documents:

* https://tools.ietf.org/html/rfc7518#page-6

Per this document, for HMAC and RSA there are clearly defined "alg"
values to be used.  This commit updates the code used to generate the
/token_key(s) documents to use the JWA algorithm names instead of the
algorithm names provided by Spring Security.  The purpose of this was to
make it so that JWK/JWS libraries coding to the specification can "just
work" instead of having to have UAA-specific code to translate the
currently provided Spring Security algorithm names to the standard JWA
algorithm names.

Fixes #282