Fix circular view path errors on login
The login endpoint should return text/html by default if no
Accept header is specified, or if the client specifies */*. However,
requests to UAA with an unexpected media type in the Accept header can
get redirected to the login page and will result in a 500 status code and
the following error:

Circular view path [login]: would dispatch back to the current
handler URL [/uaa/login] again. Check your ViewResolver setup!
(Hint: This may be the result of an unspecified view, due to
default view name generation.)

This is because the loginForHtml endpoint accepts */* but when spring
goes to render the page it can't find an appropriate view. The fix to
this is to take an additional parameter with the Accept media types
and check to see if it is compatible.