Update warning: East-West is deprecated
This intentionally removes the link to the migration documentation.
Why? Because if someone still has apps on East/West at this point,
we want them to reach out to us, not silently start migrating their
apps and risk losing them.

Please do not deploy this until March 16th.

Related tickets:
* https://github.com/18F/cg-uaa/pull/40
* https://favro.com/card/1e11108a2da81e3bd7153a7a/18F-3594