use isBlank form apache instead of isEmpty for hsqldb
The use of this method from hsqldb causes problem if jar is not available
The dependent library is not needed for usage if mysql or postgesql therefore we removed it.
However with this dependency it is a hard requirement which should be the case.

If you check the internals of isEmpty of hsqldb then it is the same if you call isBlank from apache
see also
http://technohandle.blogspot.de/2011/01/stringutilsisemtpty-vs.html