Return session_state when silent authorize results in login_required
We need to return session_state even when the user is not logged in to
prevent uaa-singular from getting stuck in an infinite loop of token
requests. Without session_state, uaa-singular's RP iframe cannot find
out from the OP iframe whether a change has occurred that would trigger
it to request a new token.

[#153568058] https://www.pivotaltracker.com/story/show/153568058