Add logging to IdTokenCreator to explain missing roles/attributes
It's possible that the role and custom_attributes claims might
disappear when building id_token in as part a refresh_token grant if the
user is coming from an identity provider configured with
storeCustomAttributes:false. This log message should hopefully help when
the inevitable support questions crop up about disappearing roles.

[#157486003] https://www.pivotaltracker.com/story/show/157486003