PollutionPreventionExtension cleans db beforeAll and afterAll
It used to only clean the db as an afterAll.

This protects tests using the PollutionPreventionExtension from
pollution in addition to preventing them from creating pollution.

The downside is that it adds ~30s to our test time (in parallel). We
would love to undo this change someday, but for the time being our tests
keep getting broken by malicious polluting evil tests.

[#164485371 #164471724]

Signed-off-by: Andrew Edstrom <aedstrom@pivotal.io>
Co-authored-by: Andrew Edstrom <aedstrom@pivotal.io>