Squash more test pollution in our unit test suite!
- Suffix the hsql db name with the gradle worker number, similar to how
we have done it with the other db types, so that hsql unit tests don't
pollute each other when they are run in parallel. Note that hsql will
automatically create these DBs when you open a connection to them, so
unlike the other db types we do not need to write our own custom code
to create them (in start_db_helper.sh)
- Refactor the db migration tests to all subclass a new common parent
class, which takes extra steps to ensure that the db is returned to the
state which other tests assume after each test (similar to the changes
that we recently made in JdbcTestBase). These db migration tests were
causing test pollution by calling flyway.clean() and leaving the db in
an unmigrated state for the next tests.
- Small change in TestUtils to create another public method to help the
db migration tests clean up after themselves

[#164124685] https://www.pivotaltracker.com/story/show/164124685