Support stopping tomcat on initialization failure
This commit provides a LifecycleListener which will shutdown the servlet
container if any servlet fails to start.

In order to make use of this utility, server.xml must be changed to
include this listener.

Do not use this utility if the UAA is deployed into a servlet container
that is running other applications -- they will be stopped as well.

[#165798297]

Signed-off-by: Andrew Wittrock <awittrock@pivotal.io>
Co-authored-by: Andrew Wittrock <awittrock@pivotal.io>