Add DisableSpringUaaSpringSecurityFilterRegistrationBean
Spring Security filters are created as beans in the XML files and then
aggregated into FilterChainProxy which is registered with Spring Boot.
However, Spring Boot registers all Beans that implement Filter with the
servlet container. This means that all the Security Fiters are executed
twice. Additionally, the Filter is going to be executed for every URL
since Boot's default is to register for every URL.

This is a post processor that marks all UAA Filter's as disabled (i.e. do
not register it with the servlet container). This way it is only executed
within FilterChainProxy