Purge expired session from in memory map
Prior to this commit, configuring the UAA to manage sessions in memory
resulted in the use of `MapSessionRepository` to manage sessions.
`MapSessionRepository` does not automatically remove expired sessions
from it's backing map.

And neither did the UAA, resulting in a memory leak.

Now, register a scheduled task to remove expired sessions. The tasks
frequency can be configured via the `servlet.session-purge-delay`
property.

[#170035178]