Minimal server.xml for UAA image
This server.xml should represent the minimal configuration necessary for
the UAA to successfully run. Tomcat behavior cannot be configured yet
and logging will go to the console for now.

Configuration of the UAA, however, is possible by providing the location
of the UAA config file via one of the supported environment variables.

[#169713142]

Signed-off-by: Jeremy Morony <jmorony@pivotal.io>