Redirect all prefixes of "/passcode" to help users
The CF CLI prints out a url when a user tries to log in with SSO:
`https://login.cloud.service.gov.uk/passcode` in Ireland and
`https://login.london/cloud.service.gov.uk/passcode` in London.

Some shells interpret the link and make it clickable… but because the
URL is so long part of `/passcode` gets trimmed onto the next line.
The shells seem to miss out on that, which is silly because the
newline is a formatting thing rather than an actual "\n".

This means that on at least half a dozen different occasions our users
have gone to `https://login.cloud.service.gov.uk/pa` or
`https://login.cloud.service.gov.uk/passco`. Those got back vague
error messages which discouraged the user from trying to solve the
problem themselves.

This commit redirect `/p`, `/pa`, `/pas`, …, `/passcod` to the
passcode page, so that things start magically working for these users.