Make the openid parameter clearer
Myself and @samze were setting up an app to go through the access token grant flow. We skimmed the docs and misread that we should add the user_attributes scope and missed the openid scope. We think this was because it is formatted more prominently so think we should do the same for the openid scope.