Allow caller with uaa.admin scope to change a user's password
The documentation for the change user password call says, for Authorization:

Access token with password.write or uaa.admin required

However the current configuration only allows the call if the caller has the
password.write scope.

This fix changes the configuration to match the documentation. It replaces:

IS_AUTHENTICATED_FULLY,scope=password.write

with the expression:

isFullyAuthenticated() and #oauth2.hasAnyScope('password.write')

The UAA custom version of hasAnyScope has an internal check for the uaa.admin
scope. It also removes the passwordResourceAuthenticationFilter, which also
insisted on a scope with the password prefix.