Fix issue where implicit grant flow fails weirdly
The implicit grant flow was returning from an authentication request
with a hash fragment that contains the state parameter.

Erroneously, this request was being rejected as the state parameter was
unable to be found as a request parameter.

We fixed this by moving the state-parameter checking to after a step
where the hash fragment parmameters are decomposed and turned into
request parameters.

[#173092312]