Fix bug involving statsd connections not being terminated.
By creating new instances of the statsd connection each time that the
task triggers, the UAA was opening a new connection to the statsd-agent.
This led to a build-up of many non-terminated connections.

Fix by adding configuration annotation to statsd configuration class so that it would
reuse the connection already created by the bean rather than creating a
new connection every time.

[#173728966]

Signed-off-by: Ryker Reed <rreed@pivotal.io>