Degraded/Limited mode tweaks to LoginPagePerformanceMockMvcTest
This is needed in continuation to changes in commits 7bc739a5bd and
31b89ccee that updated tests to use the environment rather than a status
file to signal degraded/limited mode. This has not been pushed/accepted
by upstream. By default, degraded/limited mode should be the case, so
just keep the assertion but remove the steps to tweak the status file.