Fix idpDiscoveryPageNotDisplayed_IfClientOnlyHasUAAProvider test
Upstream 75.6.0 updated the redirection code, merging of which removed
an additional check that predix extensions added for at least one IDP to
be present for the discovery email page (email.html) to be displayed. If
there is no IDP present, then the predix extension version redirects to
the username/password (password.html) page. Update upstream's changes to
bring back the check for there being at least one IDP for the
redirection to the discovery email page. This fixes the corresponding
test written specifically for this predix extension specific tweak.

This also needs an upstream unit test to be fixed to check redirection
to password page rather than email page as the test doesn't setup any
IDP.