Fix testLoginHintFallbackToLoginPage MockMvc test
Upstream v75.60.0 has changed redirection code handling of login hints
has changed. In v75.0.0, if discovery is enabled and if there is a
login hint, then the redirection is to the username/password
(password.html) page. In v75.6.0, if discovery is enabled but has not
been performed, the redirection is to the email (email.html) page [1].

Update test specifically added by predix extensions that tests
redirection to login page, to test redirection to email page rather than
password page.

[1] https://github.com/cloudfoundry/uaa/compare/75.0.0...75.6.0#diff-39666ac076d85b660eb868c942e21f648c6be235e4c829e5d2af7474b8fd617d