Fix high/medium coverity issues - Part II (#267)
* Fix possible null dereference in UaaAuthorizationEndpoint

If the host name (used as origin) can not be extracted from the redirect
URI for whatever reason, a null pointer exception will get thrown.
Instead, return a 'Bad Request' response if the host extraction fails.

* Fix possible null dereference in UaaTokenStore

Parsing authentiction from token might result in a null map that might
result in a null pointer exception. Instead, check for null, and raise
an InvalidGrantException if the parsing fails and a null is returned.

* Fix possible null dereference in RefreshTokenCreator

Move dereference of user into block that checks user for null to prevent
null pointer exceptions.

* Fix possible null dereference in IdentityProviderEndpoints

Identity provider retrieved is checked for null in DELETE handler, in
PATCH handler (but only in a patch function), and also in
ExternalOAuthProviderConfigurator:retrieve. This checking seems
unnecessary as JdbcIdentityProviderProvisioning:retrieve should throw an
EmptyResultDataAccessException rather than return null if the identity
provider isn't found. But rather than remove null checks and introduce
potential null pointer exceptions, add null checks in PUT, PATCH, and
GET handlers. If identity provider is not found, return 404 Not Found
for GET requests (by throwing EmptyResultDataAccessException), and 422
Unprocessable Entity for PUT/PATCH requests as is done for DELETE
requests.

Add/update unit tests.

* Fix possible null dereference in PasswordComparisonAuthenticator

Move check for `attr` being null to be before it gets dereferenced. Also
remove subsequent check for null for `attr`.

* Clean up unncessary code in PasswordComparisonAuthenticator

instanceof check is enough for null check too. Return value is not being
used by any (just one) calling function.

* Fix possible null dereference in IdpInitiatedLoginController

Rather than dereference a null entity descriptor, check for null and
raise an exception that will get handled by the caller.

* Fix possible null dereference in NonSnarlIdpMetadataManager

Check for metadata for null in getHostedSpName before dereferencing.
If metadata is null, throw a MetadataProviderException which will get
handled by the only calling function.

* Fix possible null dereferences in ScimUserBootstrap

AuthEvent handling might result in null dereference if fetch of SCIM
user fails when trying to update SCIM user. If SCIM user fetch fails,
and the user is null, throw an explicit RuntimeException as is done in
another handler, rather than have a NullPointerException get thrown when
the null user is dereferenced.