Fix 'UP-TO-DATE' acceptance tests not running
Acceptance tests will not be run if a previous run succeeded as gradle
doesn't see any changes to the test source files, and so assumes the
tests don't need to be run. But since the test is a 'system' test, and
the system can change (e.g. new deployment), the test needs to be run
every single time regardless of there being a code change or not.

- Update acceptanceTest task to always have tests out of date by setting
  `outputs.upToDateWhen` to be false.