Use `nurego` spring profile to enable UAA usage metering with Nurego.
Usage reporting of UAA metered resources (users and tokens) to Nurego
is configured and performed by spring metering filter [1]. Filter and,
thus, metering is enabled if `nurego` spring profile is present.
Metering is disabled and no usages are reported if `nurego` spring
profile is not present.

- Update nurego spring configuration files to use spring metering
filter if `nurego` profile is used, and to use NoOpFilter if `nurego`
profile is not used.
- Update cargo configuration for local runs to use `nurego` profile in
addition to `default` for running SpringMeteringFilterIT.
- Update taskGraph that sets context for driving integration,
acceptance, and degraded tests to also use `nurego` profile in addition
to `default`.

[1] https://github.com/predix/spring-metering-filter