Update GovCloud jenkins pipelines to explicitly checkout iam-k8s-utils
UAA checkout does not download git submodule `iam-k8s-utils` source code
in default mode it only creates empty directory for submodule.
To checkout UAA with submodule code, we need to use `--recurse-submodules`
option with `git clone` but again this submodule is a private repository
which requires another set of git credentials for accessing it.
As a workaround for fetching submodule code from this private repo,
explicitly checkout submodule code in submodule directory.

- Update UAA build pipeline to explicitly checkout git submodule
  `iam-k8s-utils` code and then compile, build and run tests.

[1] https://github.build.ge.com/predix/iam-k8s-utils/pull/1