Fix RSA public key parser (#2340)
* Fix RSA public key parser

Issue is a regression with #2284. Before the spring-security-jwt parser was used and now nimbus, but is stricter in parser of key content.
Implemented now the less strict parser from the library. We want remove spring-security-jwt in near future, thus did not used library anymore
Added test with example of YAML or DB config based oauth2 provider in parameter tokenKey. The key there does not have line break but spaces or other chars.

* changed parsed to filter out the white spaces but stay with the nimbus more generic key parser

* sonar refactorings

* refactorings