Support json key set in tokenKey
Why: OIDC IdP trust can be verified using tokenKeyUrl or tokenKey
In case of tokenKeyUrl the jwks_uri is called to fetch all current keys in order to verify the JWT signature.
In case of tokenKey there is currently only a single key possible. If the IdP has a dynamic key rotation, then trust breaks.