feature: Activate PKCE by default in OIDC client request
PKCE for OIDC/OPENID clients is available now since years in UAA but currently only on/active for public usage
Add now an own option to have it on also for all other authentications. Default on, but if a client does not want it, you can switch it off.
See https://oauth.net/2/pkce/, PKCE is not only for public but it is use in this scenarios it most frameworks (spring and golang)

https://oauth.net/2.1/