fix: do not default a missing secret to an empty one
If client has no secret set, then configuration should have no secret

Before: a missing secret was set as empty and this empty secret can be used even with client credential.
After: a missing secret is null in DB layer so that you cannot do secret based login. There are other authentications so that the client
is still usable.