Update account menu for keyboard accessibility
Previously, the dropdown account menu in the top right could only be
used by clicking on it using a mouse. People who use screen readers,
such as those who are blind or have low-vision, require the ability to
navigate and interact using keyboard-only. These users cannot use a
mouse for user interaction.

Now, the account menu can be interacted with using keyboard. It uses the
`aria-expanded` and `aria-controls` to provide better feedback to screen
readers about whether the menu is opened or closed.