fix flaky test
This class and many nested test classes
The result is, that jku_test is deleted and recreated in parallel and
this leads to inconsistency which can be solved with the retry.
We have this retry on other places already.

Therefore add it now here again. In real such issues wont happen therefore we maybe
should solve this with creating own clients for all nested tests

Most failing test seen validExpClaim