fix: User OIDC standard claims, if mapping is ambiguous
With https://github.com/cloudfoundry/uaa/pull/1925 there was a fix already, however using custom claims can lead to error situations.
With this fix, that error situations are prevented if standard claims in id_token, e.g. name is an array with different values,
but family_name and given_name are also in the token as string.
This fix is for OIDC IdP integration, therefore it makes sense to use as fallback also OIDC standard claims.