fix client admin rest call with empty secret
Empty client secret is allowed if client secret policy allows it.
Error says 1 char is needed, this is wrong

Why 2 validators, distinguish between user and client policy validation.

For users 1 minimum char for a password might be ok, for a client definitely not.
The empty secret should be treated as public, but we cannot disallow it.

Therefore, this here is a fix for a regression introduced with 76.22.0.