feature: add client_auth_method=none into tokens for clients with empty secret (#2504)
* feature: add client_auth_method=none into tokens for clients with empty secret

e.g. cf client.
This allows clients to see, whats really used.
According to https://oauth.net/2.1/ the empty secret can be treated as public usage

* added tests

* refactor

* refactor back

* more tests

* revert defaults

* allow empty secret if secret policy minLength == 0

* Add client_auth_method=empty and add addtional tests

empty is a valid authentication with empty secret, however it is no OAuth2 standard

* refactor

* feature: add client_auth_method=none into tokens for clients with empty secret

Internally use empty, but write none into tokens. Externally use always none, but internally empty to distinguish in the flow.
The refresh flow with an empty secret is something we allow for now.

* refactor: test names & fix typos

- such that it's clearer what the test cases are

* rebase

* comment about internal and external client auth

---------

Co-authored-by: Peter Chen <peterch@vmware.com>