reduce the defaults
Return the oidc standard claim if the mapped claim is not in token.
This is useful needed if the mapping is not done or not correct.
If the mapping is available this is used.

The reality has shown, that many operators do not create a mapping at all or
correct mapping, but with OIDC the mapping for email, given_name, family_name is not
needed because they are there with scope profile.

See test getUser_doesNotReturnNullWhenIdTokenMappingIsNotAvailableButAlsoOidcStandardClaims