Refactor: Load UserConfig as bean
Goal: identityZoneConfigurationBootstrap should not get
to many attributes, but load the user config itself and
allow there to have extra options

Example: login.checkOriginEnabled is not yet possible to
set in CF global because it is / was hidden in the
identityZoneConfigurationBootstrap bean. With UserConfig
bean allow all options to be configured.

Open: uaa-release should be adopted as well