fix: Integration test bootstrap fixed (#2959)
* fix: Integration test bootstrap fixed

These tests need a bootstraping which is not in the test itself but externally.
This means, these tests cannot be executed standalone, which is bad. If another test runs before, then these tests can run as well, therefore it has not been found before. Normally we run all tests in a suite.

* found another test