Fix regression in identity-provider endpoint (#2962)
* Fix regression in identity-provider endpoint

Issue:
If existing entries in identity-provider with new external_key the field is null, which is expected.
If external_key is null, this must not overwrite the issuer in rest endpoint, but it does

For SAML there is no issue, because here the entityId is really new in REST output and in DB.
For OIDC and OAuth2 the issuer was used in REST already and there was no check before overwrite it from external_key.

* review

* add case if issuer is null from config, allowed for oauth2 IdP

* spelling

* revert the logic of external key, stay with issuer

* set entityId on update

* test coverage

Co-authored-by: Peter Chen <peter-h.chen@broadcom.com>