fix saml validation and allow again encryption (only) assertions
Tested side effect: In former UAAs the option

wantAssertionSigned: false

was too dangerous because you could end up in a validation
without any crypto part.

Now you can set in UAA default zone wantAssertionSigned: false because there is an extra check of situations without crypto part.
But if an assertion is encrypted it we can read it, then we dont need the signature because the encrypted is good enough (or even better than signature only).